ui優化，fix bug,去掉使用AsyncTask，改用Thread和Executors

+++ app/src/main/java/com/example/moneymanager/addBalanceActivity.java
@@ -21,6 +21,8 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 public class addBalanceActivity extends AppCompatActivity {
 
@@ -55,35 +57,28 @@ public class addBalanceActivity extends AppCompatActivity {
         expenseDatabase = ExpenseDatabase.getInstance(this);
 
 
-        AsyncTask.execute(new Runnable() {
+        new Thread(new Runnable() {
             @Override
             public void run() {
-                if (expenseID!=-1)
-                {
+                if (expenseID!=-1) {
                     expense = expenseDatabase.expenseDao().getExpenseById(expenseID);
-
-                    if (expense!=null) {
-                        setText(nameEditText,balanceEditText,addBalanceDateBtn,expense);
-                    }
+                    if (expense!=null) setText(nameEditText,balanceEditText,addBalanceDateBtn,expense);
                 }
             }
-        });
+        }).start();
 
         addBalanceDelBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if (expenseID==-1)
-                    Toast.makeText(getApplicationContext(),"沒有資料刪除！！",Toast.LENGTH_SHORT).show();
+                if (expenseID==-1) Toast.makeText(getApplicationContext(),"沒有資料刪除！！",Toast.LENGTH_SHORT).show();
                 else{
-                    AsyncTask.execute(new Runnable() {
+                    new Thread(new Runnable() {
                         @Override
                         public void run() {
                             Expense expense1 = expenseDatabase.expenseDao().getExpenseById(expenseID);
-                            if (expense1!=null){
-                                expenseDatabase.expenseDao().deleteExpense(expense1);
-                            }
+                            if (expense1!=null) expenseDatabase.expenseDao().deleteExpense(expense1);
                         }
-                    });
+                    }).start();
                 }
                 Intent intent = new Intent(addBalanceActivity.this,MainActivity2.class);
                 intent.putExtra("Reload",1);
@@ -163,9 +158,10 @@ public class addBalanceActivity extends AppCompatActivity {
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM", Locale.getDefault());
         String selectedDateString = sdf.format(date);//addBalanceDateBtn選中的日期
 
-        AsyncTask<Void, Void, Void> saveExpenseTask = new AsyncTask<Void, Void, Void>() {
+        ExecutorService saveExpenseTask = Executors.newSingleThreadExecutor();
+        saveExpenseTask.execute(new Runnable() {
             @Override
-            protected Void doInBackground(Void... voids) {
+            public void run() {
                 ExpenseDao expenseDao = expenseDatabase.expenseDao();
 
                 // 检查数据库中是否存在特定类别的数据
@@ -173,8 +169,7 @@ public class addBalanceActivity extends AppCompatActivity {
 
                 if (expenses != null && expenses.size() > 0) {
                     // 如果存在，执行更新操作
-                    for (int i=0;i<expenses.size();i++)
-                    {
+                    for (int i=0;i<expenses.size();i++) {
                         Expense expense = expenses.get(i);
 
                         String expenseDate = sdf.format(expense.getDate());
@@ -188,24 +183,17 @@ public class addBalanceActivity extends AppCompatActivity {
                             Expense expense1 = new Expense(Double.parseDouble(amount), name, date, category);
                             expenseDao.insertExpense(expense1);
                         }
-
                     }
-                } else {
+                }
+                else {
                     // 如果不存在，执行插入操作
                     Expense expense = new Expense(Double.parseDouble(amount), name, date, category);
                     expenseDao.insertExpense(expense);
                 }
 
-                return null;
             }
-
-            @Override
-            protected void onPostExecute(Void aVoid) {
-                super.onPostExecute(aVoid);
-            }
-        };
-
-        saveExpenseTask.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
+        });
+        saveExpenseTask.shutdown();
     }
